// apps/web/src/types/supabase.ts
export type Player = {
  id: number;
  name: string;
  email: string;
  positions: string[];
  preferredPosition: string;
  skillRating: number;
  availability: boolean;
  gender: "male" | "female" | "other";
  matchesPlayed: number;
  isPublic: boolean;
};